+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## 4.14.0
+
+* Adds support for configuring file URL access permissions. See
+  `AndroidWebViewController.setAllowFileAccessFromFileURLs` and
+  `AndroidWebViewController.setAllowUniversalAccessFromFileURLs`.
 * Adds support for configuring Web Authentication in `AndroidWebViewController` with `setWebAuthenticationSupport` to enable Passkey and other related Authentication.
 
 ## 4.13.0